Internal

The following is internal, i.e. not public API, and might change at any point.

Index

OhMyThreads.Tools.nthtid

OhMyThreads.Tools.taskid

References

OhMyThreads.Tools.nthtid — Method

Returns the thread id of the nth Julia thread in the :default threadpool.

OhMyThreads.Tools.taskid — Method

taskid() :: UInt

Return a UInt identifier for the current running Task. This identifier will be unique so long as references to the task it came from still exist.