Public API

Index

OhMyThreads.tcollect

OhMyThreads.tforeach

OhMyThreads.tmap

OhMyThreads.tmap!

OhMyThreads.tmapreduce

OhMyThreads.treduce

OhMyThreads.treducemap

OhMyThreads.Schedulers.DynamicScheduler

OhMyThreads.Schedulers.GreedyScheduler

OhMyThreads.Schedulers.Scheduler

OhMyThreads.Schedulers.SpawnAllScheduler

OhMyThreads.Schedulers.StaticScheduler

Exported

Functions

OhMyThreads.tmapreduce — Function

tmapreduce(f, op, A::AbstractArray...;
           [scheduler::Scheduler = DynamicScheduler()],
           [outputtype::Type = Any],
           [init])

A multithreaded function like Base.mapreduce. Perform a reduction over A, applying a single-argument function f to each element, and then combining them with the two-argument function op.

Note that op must be an associative function, in the sense that op(a, op(b, c)) ≈ op(op(a, b), c). If op is not (approximately) associative, you will get undefined results.

Example:

tmapreduce(√, +, [1, 2, 3, 4, 5])

is the parallelized version of sum(√, [1, 2, 3, 4, 5]) in the form

(√1 + √2) + (√3 + √4) + √5

Keyword arguments:

scheduler::Scheduler (default DynamicScheduler()): determines how the computation is divided into parallel tasks and how these are scheduled. See Scheduler for more information.

outputtype::Type (default Any): will work as the asserted output type of parallel calculations. We use StableTasks.jl to make setting this option unnecessary, but if you experience problems with type stability, you may be able to recover it with this keyword argument.

init: forwarded to mapreduce for the task-local sequential parts of the calculation.

OhMyThreads.treduce — Function

treduce(op, A::AbstractArray...;
        [scheduler::Scheduler = DynamicScheduler()],
        [outputtype::Type = Any],
        [init])

A multithreaded function like Base.reduce. Perform a reduction over A using the two-argument function op.

Note that op must be an associative function, in the sense that op(a, op(b, c)) ≈ op(op(a, b), c). If op is not (approximately) associative, you will get undefined results.

Example:

treduce(+, [1, 2, 3, 4, 5])

is the parallelized version of sum([1, 2, 3, 4, 5]) in the form

(1 + 2) + (3 + 4) + 5

Keyword arguments:

scheduler::Scheduler (default DynamicScheduler()): determines how the computation is divided into parallel tasks and how these are scheduled. See Scheduler for more information.

outputtype::Type (default Any): will work as the asserted output type of parallel calculations. We use StableTasks.jl to make setting this option unnecessary, but if you experience problems with type stability, you may be able to recover it with this keyword argument.

init: forwarded to mapreduce for the task-local sequential parts of the calculation.

OhMyThreads.tmap — Function

tmap(f, [OutputElementType], A::AbstractArray...;
     [schedule::Scheduler = DynamicScheduler()])

A multithreaded function like Base.map. Create a new container similar to A and fills it in parallel such that the ith element is equal to f(A[i]).

The optional argument OutputElementType will select a specific element type for the returned container, and will generally incur fewer allocations than the version where OutputElementType is not specified.

Example:

tmap(sin, 1:10)

Keyword arguments:

scheduler::Scheduler (default DynamicScheduler()): determines how the computation is divided into parallel tasks and how these are scheduled. See Scheduler for more information.

OhMyThreads.tmap! — Function

tmap!(f, out, A::AbstractArray...;
      [schedule::Scheduler = DynamicScheduler()])

A multithreaded function like Base.map!. In parallel on multiple tasks, this function assigns each element of out[i] = f(A[i]) for each index i of A and out.

Keyword arguments:

scheduler::Scheduler (default DynamicScheduler()): determines how the computation is divided into parallel tasks and how these are scheduled. See Scheduler for more information.

OhMyThreads.tforeach — Function

tforeach(f, A::AbstractArray...;
         [schedule::Scheduler = DynamicScheduler()]) :: Nothing

A multithreaded function like Base.foreach. Apply f to each element of A on multiple parallel tasks, and return nothing. I.e. it is the parallel equivalent of

for x in A
    f(x)
end

Example:

tforeach(1:10) do i
    println(i^2)
end

Keyword arguments:

scheduler::Scheduler (default DynamicScheduler()): determines how the computation is divided into parallel tasks and how these are scheduled. See Scheduler for more information.

OhMyThreads.tcollect — Function

tcollect([OutputElementType], gen::Union{AbstractArray, Generator{<:AbstractArray}};
         [schedule::Scheduler = DynamicScheduler()])

A multithreaded function like Base.collect. Essentially just calls tmap on the generator function and inputs.

The optional argument OutputElementType will select a specific element type for the returned container, and will generally incur fewer allocations than the version where OutputElementType is not specified.

Example:

tcollect(sin(i) for i in 1:10)

Keyword arguments:

scheduler::Scheduler (default DynamicScheduler()): determines how the computation is divided into parallel tasks and how these are scheduled. See Scheduler for more information.

OhMyThreads.treducemap — Function

treducemap(op, f, A::AbstractArray...;
           [scheduler::Scheduler = DynamicScheduler()],
           [outputtype::Type = Any],
           [init])

Like tmapreduce except the order of the f and op arguments are switched. This is sometimes convenient with do-block notation. Perform a reduction over A, applying a single-argument function f to each element, and then combining them with the two-argument function op.

Note that op must be an associative function, in the sense that op(a, op(b, c)) ≈ op(op(a, b), c). If op is not (approximately) associative, you will get undefined results.

Example:

tmapreduce(√, +, [1, 2, 3, 4, 5])

is the parallelized version of sum(√, [1, 2, 3, 4, 5]) in the form

(√1 + √2) + (√3 + √4) + √5

Keyword arguments:

scheduler::Scheduler (default DynamicScheduler()): determines how the computation is divided into parallel tasks and how these are scheduled. See Scheduler for more information.

outputtype::Type (default Any): will work as the asserted output type of parallel calculations. We use StableTasks.jl to make setting this option unnecessary, but if you experience problems with type stability, you may be able to recover it with this keyword argument.

init: forwarded to mapreduce for the task-local sequential parts of the calculation.

Schedulers

OhMyThreads.Schedulers.Scheduler — Type

Supertype for all available schedulers:

DynamicScheduler: default dynamic scheduler

StaticScheduler: low-overhead static scheduler

GreedyScheduler: greedy load-balancing scheduler

SpawnAllScheduler: @spawn one task per element

OhMyThreads.Schedulers.DynamicScheduler — Type

The default dynamic scheduler. Divides the given collection into chunks and then spawns a task per chunk to perform the requested operation in parallel. The tasks are assigned to threads by Julia's dynamic scheduler and are non-sticky, that is, they can migrate between threads.

Generally preferred since it is flexible, can provide load balancing, and is composable with other multithreaded code.

Keyword arguments:

nchunks::Int (default 2 * nthreads(threadpool)):

Determines the number of chunks (and thus also the number of parallel tasks).

Increasing nchunks can help with load balancing, but at the expense of creating more overhead. For nchunks <= nthreads() there are not enough chunks for any load balancing.

Setting nchunks < nthreads() is an effective way to use only a subset of the available threads.

split::Symbol (default :batch):

Determines how the collection is divided into chunks. By default, each chunk consists of contiguous elements.

See ChunkSplitters.jl for more details and available options.

threadpool::Symbol (default :default):

Possible options are :default and :interactive.

The high-priority pool :interactive should be used very carefully since tasks on this threadpool should not be allowed to run for a long time without yielding as it can interfere with heartbeat processes.

OhMyThreads.Schedulers.StaticScheduler — Type

A static low-overhead scheduler. Divides the given collection into chunks and then spawns a task per chunk to perform the requested operation in parallel. The tasks are statically assigned to threads up front and are made sticky, that is, they are guaranteed to stay on the assigned threads (no task migration).

Can sometimes be more performant than DynamicScheduler when the workload is (close to) uniform and, because of the lower overhead, for small workloads. Isn't well composable with other multithreaded code though.

Keyword arguments:

nchunks::Int (default nthreads()):

Determines the number of chunks (and thus also the number of parallel tasks).

Setting nchunks < nthreads() is an effective way to use only a subset of the available threads.

Currently, nchunks > nthreads() isn't officialy supported but, for now, will fall back to nchunks = nthreads().

split::Symbol (default :batch):

Determines how the collection is divided into chunks. By default, each chunk consists of contiguous elements.

See ChunkSplitters.jl for more details and available options.

OhMyThreads.Schedulers.GreedyScheduler — Type

A greedy dynamic scheduler. The elements of the collection are first put into a Channel and then dynamic, non-sticky tasks are spawned to process channel content in parallel.

Note that elements are processed in a non-deterministic order, and thus a potential reducing function must be commutative in addition to being associative, or you could get incorrect results!

Can be good choice for load-balancing slower, uneven computations, but does carry some additional overhead.

Keyword arguments:

ntasks::Int (default nthreads()):

Determines the number of parallel tasks to be spawned.

Setting nchunks < nthreads() is an effective way to use only a subset of the available threads.

OhMyThreads.Schedulers.SpawnAllScheduler — Type

A scheduler that spawns a task per element (i.e. there is no internal chunking) to perform the requested operation in parallel. The tasks are assigned to threads by Julia's dynamic scheduler and are non-sticky, that is, they can migrate between threads.

Note that, depending on the input, this scheduler might spawn many(!) tasks and can be very costly!

Essentially the same as DynamicScheduler(; nchunks=nelements), but with a simpler, potentially more efficient implementation.

Keyword arguments:

threadpool::Symbol (default :default):

Possible options are :default and :interactive.

The high-priority pool :interactive should be used very carefully since tasks on this threadpool should not be allowed to run for a long time without yielding as it can interfere with heartbeat processes.

Non-Exported

OhMyThreads.@spawn	see StableTasks.jl

OhMyThreads.@spawnat	see StableTasks.jl

OhMyThreads.chunks	see ChunkSplitters.jl

OhMyThreads.TaskLocalValue	see TaskLocalValues.jl